backend/requirements.txt
fastapi==0.135.3
uvicorn==0.44.0
pymongo[srv]==4.6.1
python-dotenv==1.2.2
passlib[bcrypt]==1.7.4
python-jose==3.5.0
pydantic==2.13.0
google-genai==1.73.0
